+++ packages/theme/demo/changelog.md
@@ -7,6 +7,15 @@ tag:
 
 # Changelog
 
+升级主题，原项目执行如下指令即可
+```sh
+pnpm up @sugarat/theme
+```
+
+## 0.1.10 (2023/03/12)
+- feat: 支持全文搜索（基于 [pagefind](https://pagefind.app/) 实现）
+  - `search: 'pagefind'`
+
 ## 0.1.9 (2023/02/24)
 ### Patch Changes
 - feat: 支持自定义推荐文章的展示顺序，详见[frontmatter.recommend](./config/frontmatter.md#recommend)